+++ CHANGELOG.md
@@ -2,11 +2,34 @@
 
 ## v2.0.0
 
-* **不兼容改动** 升级 JavaScript SDK 到 [leancloud-storage 2.1.0](https://github.com/leancloud/javascript-sdk/releases/tag/v2.0.0)（请检查 JS SDK 2.0 引入的不兼容改动）
-* **不兼容改动** 仅兼容至 Node.js 4.4 以上版本
-* **不兼容改动** 重复定义云函数或 Hook 时会抛出一个异常
-* **不兼容改动** 移除 `AV.Cloud.httpRequest`（请改用 `request` 模块）
-* **不兼容改动** 移除 `app.use(AV.Cloud)` 的用法（请改用 `app.use(AV.express())`）
+- **不兼容改动** 升级 JavaScript SDK 到 [leancloud-storage 2.1.0](https://github.com/leancloud/javascript-sdk/releases/tag/v2.0.0)（请检查 JS SDK 2.0 引入的不兼容改动）
+- **不兼容改动** 仅兼容至 Node.js 4.4 以上版本
+- **不兼容改动** 重复定义云函数或 Hook 时会抛出一个异常
+- **不兼容改动** 移除 `AV.Cloud.httpRequest`（请改用 `request` 模块）
+- **不兼容改动** 移除 `app.use(AV.Cloud)` 的用法（请改用 `app.use(AV.express())`）
+- **不兼容改动** 移除基于 [domain](https://nodejs.org/api/domain.html)，开发者需要自行捕捉云函数异步代码中的异常。
+
+新增云函数和 Class Hook 的 Promise 模式，会使用 Promise 的值作为响应内容。如果在 Promise 中抛了使用新增的 `AV.Cloud.Error` 构造的异常则作为错误返回给客户端，`AV.Cloud.Error` 的第二个参数可以指定 HTTP Status Code 和 Error Code（`AV.Cloud.Error('posts is empty', {status: 422, code: 422})`）；如果抛出了其他错误类型则视作服务器端错误，返回 500 响应并打印错误到标准输出。
+
+```javascript
+AV.Cloud.define(function(request) {
+  return new AV.Query('Post').find().then( posts => {
+    if (posts.length > 0) {
+      return posts[0];
+    } else {
+      throw new AV.Cloud.Error('posts is empty');
+    }
+  });
+});
+```
+
+如果传入 `AV.Cloud.define` 的函数有两个参数（`request` 和 `response`）则继续兼容原定义方式，需要使用 `response.success()` 发送响应。我们会继续兼容这种用法到下一个大版本，希望用户尽快迁移到 Promise 风格的云函数上。
+
+## v1.2.4
+
+暂时地锁定 leancloud-storage 的版本，控制 [1.5.5](https://github.com/leancloud/javascript-sdk/releases/tag/v1.5.5) 中 disableCurrentUser 变动的影响。
+
+- 锁定 leancloud-storage 的版本到 1.5.4
 
 ## v1.2.3